backend/requirements.txt
fastapi
uvicorn[standard]

motor
pymongo

python-jose[cryptography]
passlib[bcrypt]

pydantic
pydantic-settings
python-dotenv

python-multipart

reportlab
qrcode[pil]
pillow
email-validator
